+++ test/unit/fuzz-tools.test.js
@@ -316,6 +316,7 @@ import * as jnetV336 from '../../lib/jnet-v336.js';
 import * as outerbridgeV337 from '../../lib/outerbridge-v337.js';
 import * as icrsV338 from '../../lib/icrs-v338.js';
 import * as cormackLehaneV339 from '../../lib/cormack-lehane-v339.js';
+import * as clarkLevelV340 from '../../lib/clark-level-v340.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -577,6 +578,7 @@ const MODULES = {
   'outerbridge-v337.js': outerbridgeV337,
   'icrs-v338.js': icrsV338,
   'cormack-lehane-v339.js': cormackLehaneV339,
+  'clark-level-v340.js': clarkLevelV340,
   'tb-testing.js': tbTesting,
 };